microcontroller update
i have made sure that the data for settings are cheked before used to make sure that the information is correct. i have also made it so if the user chooses to turn animated leds on and off, this is also handled correctly. i dont think there is any other work that needs to be done on the microcontroller.